requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
docling
lancedb
streamlit
tiktoken
streamlit
python-dotenv
langchain-docling
langchain-core
langchain-huggingface
langchain-milvus
langchain-text-splitters
pymilvus
docling-core
sentence-transformers